repl

+++ utils/pack.cpp
@@ -19,14 +19,14 @@ void naruto::utils::Pack::serialize(const ::google::protobuf::Message &msg, uint
     msg.SerializeToOstream(out);
 }
 
-uint64_t naruto::utils::Pack::deSerialize(naruto::utils::Bytes &pack, ::google::protobuf::Message &msg) {
+uint16_t naruto::utils::Pack::deSerialize(naruto::utils::Bytes &pack, ::google::protobuf::Message &msg) {
     auto len = pack.getInt(PACK_HEAD_SIZE_INDEX);
     auto type =  pack.getShort(PACK_HEAD_TYPE_INDEX);
     msg.ParseFromArray(&pack.data()[PACK_HEAD_LEN], len - PACK_HEAD_LEN);
     return type;
 }
 
-uint64_t naruto::utils::Pack::deSerialize(const unsigned char *data, size_t n, google::protobuf::Message &msg) {
+uint16_t naruto::utils::Pack::deSerialize(const unsigned char *data, size_t n, google::protobuf::Message &msg) {
     auto len = (uint32_t)data[PACK_HEAD_SIZE_INDEX];
     auto type = (uint16_t)data[PACK_HEAD_TYPE_INDEX];
     msg.ParseFromArray(&data[PACK_HEAD_LEN], len - PACK_HEAD_LEN);
